services/template-service/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy[asyncio]
asyncpg
redis[asyncio]
pydantic
python-dotenv
alembic
python-multipart
pytest
psycopg
psycopg[binary,pool]
psycopg2